refactor: dto product 로 변경
