+++ src/main/java/cart/dto/product/ProductResponse.java
@@ -1,4 +1,4 @@
-package cart.dto;
+package cart.dto.product;
 
 import cart.domain.product.Product;
 import java.util.Objects;
