+++ src/test/java/cart/dto/product/ResponseFixture.java
@@ -1,4 +1,4 @@
-package cart.dto;
+package cart.dto.product;
 
 @SuppressWarnings("SpellCheckingInspection")
 public class ResponseFixture {
